id -> short now int32 + color selection code updated

+++ src/common/util.h
@@ -39,7 +39,7 @@ std::string serializeRotation(glm::quat rotation);
 float maxvalue(float x, float y, float z);
 int maxvalue(int x, int y, int z);
 
-typedef short objid;
+typedef int32_t objid;
 objid getUniqueObjId();
 std::string getUniqueObjectName();
