+++ src/main_api.cpp
@@ -9,7 +9,7 @@ extern GameObject defaultCamera;
 extern std::map<unsigned int, Mesh> fontMeshes;
 extern unsigned int uiShaderProgram;
 extern float initialTime;
-extern std::vector<short> playbacksToRemove;
+extern std::vector<int32_t> playbacksToRemove;
 extern std::queue<std::string> channelMessages;
 
 extern float now;
@@ -25,7 +25,7 @@ NetworkPacket toNetworkPacket(UdpPacket& packet){
   return netpacket;
 }
 
-void setActiveCamera(short cameraId){
+void setActiveCamera(int32_t cameraId){
   auto cameraIndexs = getGameObjectsIndex<GameObjectCamera>(world.objectMapping);
   if (! (std::find(cameraIndexs.begin(), cameraIndexs.end(), cameraId) != cameraIndexs.end())){
     std::cout << "index: " << cameraId << " is not a valid index" << std::endl;
@@ -43,7 +43,7 @@ void nextCamera(){
   }
 
   state.activeCamera = (state.activeCamera + 1) % cameraIndexs.size();
-  short activeCameraId = cameraIndexs.at(state.activeCamera);
+  int32_t activeCameraId = cameraIndexs.at(state.activeCamera);
   setActiveCamera(activeCameraId);
 }
 void moveCamera(glm::vec3 offset){
@@ -53,37 +53,37 @@ void rotateCamera(float xoffset, float yoffset){
   defaultCamera.transformation.rotation = setFrontDelta(defaultCamera.transformation.rotation, xoffset, yoffset, 0, 0.1);
 }
 
-void applyImpulse(short index, glm::vec3 impulse){
+void applyImpulse(int32_t index, glm::vec3 impulse){
   applyImpulse(world.rigidbodys.at(index), impulse);
 }
-void applyImpulseRel(short index, glm::vec3 impulse){
+void applyImpulseRel(int32_t index, glm::vec3 impulse){
   glm::vec3 relativeImpulse = calculateRelativeOffset(getGameObject(world, index).transformation.rotation, impulse, true);
   applyImpulse(world.rigidbodys.at(index), relativeImpulse);
 }
 
-void clearImpulse(short index){
+void clearImpulse(int32_t index){
   clearImpulse(world.rigidbodys.at(index));
 }
 
-void loadScriptFromWorld(std::string script, short id){
+void loadScriptFromWorld(std::string script, int32_t id){
   auto name = getGameObject(world, id).name;
   std::cout << "gameobj: " << name << " wants to load script: (" << script << ")" << std::endl;
   loadScript(script, id, bootStrapperMode);
 }
-short loadScene(std::string sceneFile){
+int32_t loadScene(std::string sceneFile){
   std::cout << "INFO: SCENE LOADING: loading " << sceneFile << std::endl;
   return addSceneToWorld(world, sceneFile, interface);
 }
-short loadSceneObj(std::string sceneFile, short sceneId){
+int32_t loadSceneObj(std::string sceneFile, int32_t sceneId){
   std::cout << "INFO: SCENE LOADING: loading subscene" << sceneFile << std::endl;
 }
 
-short loadSceneData(std::string sceneData, objid sceneId){
+int32_t loadSceneData(std::string sceneData, objid sceneId){
   std::cout << "INFO: SCENE LOADING: loading from scene data" << std::endl;
   return addSceneToWorldFromData(world, sceneId, sceneData, interface);
 }
 
-void unloadScene(short sceneId){  
+void unloadScene(int32_t sceneId){  
   std::cout << "INFO: SCENE LOADING: unloading " << sceneId << std::endl;
   removeSceneFromWorld(world, sceneId, interface);
 }
@@ -99,15 +99,15 @@ void saveScene(bool includeIds){
   saveFile(fileToSave, serializeScene(world, id, includeIds));
 }
 
-std::vector<short> listScenes(){
-  std::vector<short> sceneIds;
+std::vector<int32_t> listScenes(){
+  std::vector<int32_t> sceneIds;
   for (auto &[id, _] : world.scenes){
     sceneIds.push_back(id);
   }
   return sceneIds;
 }
 
-void sendLoadScene(short id){
+void sendLoadScene(int32_t id){
   if (!bootStrapperMode){
     std::cout << "ERROR: cannot send load scene in not-server mode" << std::endl;
     assert(false);
@@ -130,7 +130,7 @@ std::optional<objid> getGameObjectByName(std::string name){    // @todo : odd be
   return getGameObjectByName(world, name);
 }
 
-std::vector<short> getObjectsByType(std::string type){
+std::vector<int32_t> getObjectsByType(std::string type){
   if (type == "mesh"){
     std::vector indexes = getGameObjectsIndex<GameObjectMesh>(world.objectMapping);
     return indexes;
@@ -140,41 +140,41 @@ std::vector<short> getObjectsByType(std::string type){
   }
   return getGameObjectsIndex(world.objectMapping);
 }
-std::string getGameObjectName(short index){
+std::string getGameObjectName(int32_t index){
   return getGameObject(world, index).name;
 }
 
-std::map<std::string, std::string> getGameObjectAttr(short id){
+std::map<std::string, std::string> getGameObjectAttr(int32_t id){
   return getAttributes(world, id);
 }
-void setGameObjectAttr(short id, std::map<std::string, std::string> attr){
+void setGameObjectAttr(int32_t id, std::map<std::string, std::string> attr){
   setAttributes(world, id, attr);
 }
 
-glm::vec3 getGameObjectPosition(short index, bool isWorld){
+glm::vec3 getGameObjectPosition(int32_t index, bool isWorld){
   if (isWorld){
     return fullTransformation(world, index).position;
   }
   return getGameObject(world, index).transformation.position;
 }
-void setGameObjectPosition(short index, glm::vec3 pos){
+void setGameObjectPosition(int32_t index, glm::vec3 pos){
   physicsTranslateSet(world, index, pos);
 }
-void setGameObjectPositionRelative(short index, float x, float y, float z, bool xzPlaneOnly){
+void setGameObjectPositionRelative(int32_t index, float x, float y, float z, bool xzPlaneOnly){
   auto transformation = getGameObject(world, index).transformation;
   glm::vec3 pos = moveRelative(transformation.position, transformation.rotation, glm::vec3(x, y, z), xzPlaneOnly);
   physicsTranslateSet(world, index, pos);
 }
-glm::vec3 getGameObjectScale(short index){
+glm::vec3 getGameObjectScale(int32_t index){
   return getGameObject(world, index).transformation.scale;
 }
-void setGameObjectScale(short index, glm::vec3 scale){
+void setGameObjectScale(int32_t index, glm::vec3 scale){
   physicsScaleSet(world, index, scale);
 }
-void setGameObjectRotation(short index, glm::quat rotation){
+void setGameObjectRotation(int32_t index, glm::quat rotation){
   physicsRotateSet(world, index, rotation);
 }
-glm::quat getGameObjectRotation(short index, bool isWorld){
+glm::quat getGameObjectRotation(int32_t index, bool isWorld){
   if (isWorld){
     return fullTransformation(world, index).rotation;
   }
@@ -185,7 +185,7 @@ void setSelectionMode(bool enabled){
   state.isSelectionMode = enabled;
 }
 
-short makeObject(std::string serializedobj, objid id, bool useObjId, objid sceneId, bool useSceneId){
+int32_t makeObject(std::string serializedobj, objid id, bool useObjId, objid sceneId, bool useSceneId){
   return addObjectToScene(world, useSceneId ? sceneId : world.scenes.begin() -> first, serializedobj, id, useObjId, interface);
 }
 objid makeObjectAttr(std::string name, std::map<std::string, std::string> stringAttributes, std::map<std::string, double> numAttributes, std::map<std::string, glm::vec3> vecAttributes){
@@ -199,7 +199,7 @@ objid makeObjectAttr(std::string name, std::map<std::string, std::string> string
   return addObjectToScene(world, world.scenes.begin() -> first, name, attributes, interface);
 }
 
-void removeObjectById(short id){
+void removeObjectById(int32_t id){
   removeObjectFromScene(world, id, interface);
 }
 
@@ -207,7 +207,7 @@ void drawText(std::string word, float left, float top, unsigned int fontSize){
   drawWords(uiShaderProgram, fontMeshes, word, left, top, fontSize);
 }
 
-std::vector<std::string> listAnimations(short id){
+std::vector<std::string> listAnimations(int32_t id){
   std::vector<std::string> animationNames;
   auto groupId = getGroupId(world, id);
   if (world.animations.find(groupId) == world.animations.end()){
@@ -219,7 +219,7 @@ std::vector<std::string> listAnimations(short id){
   }
   return animationNames;
 }
-Animation getAnimation(World& world, short groupId, std::string animationToPlay){  
+Animation getAnimation(World& world, int32_t groupId, std::string animationToPlay){  
   Animation noAnimation { };
   for (auto animation :  world.animations.at(groupId)){
     if (animation.name == animationToPlay){
@@ -231,7 +231,7 @@ Animation getAnimation(World& world, short groupId, std::string animationToPlay)
   return  noAnimation;  // @todo probably use optional here.
 }
 
-void addAnimation(AnimationState& animationState, short groupId, std::string animationToPlay){
+void addAnimation(AnimationState& animationState, int32_t groupId, std::string animationToPlay){
   auto animation = getAnimation(world, groupId, animationToPlay);
   auto meshNameToMeshes = getMeshesForGroupId(world, groupId);  
   TimePlayback playback(
@@ -247,7 +247,7 @@ void addAnimation(AnimationState& animationState, short groupId, std::string ani
   animationState.playbacks[groupId] = playback;
 }
 
-void playAnimation(short id, std::string animationToPlay){
+void playAnimation(int32_t id, std::string animationToPlay){
   auto groupId =  world.scenes.at(world.idToScene.at(id)).idToGameObjectsH.at(id).groupId;
   if (animations.playbacks.find(groupId) != animations.playbacks.end()){
     animations.playbacks.erase(groupId);
@@ -272,11 +272,11 @@ void sendNotifyMessage(std::string message){
   channelMessages.push(message);
 }
 
-void attachToRail(short id, std::string rail){
+void attachToRail(int32_t id, std::string rail){
   addEntity(world.rails, id, rail);
 }
 
-void unattachFromRail(short id){
+void unattachFromRail(int32_t id){
   removeEntity(world.rails, id);
 }
